requirements.txt
numpy==1.12.1
Pillow==5.1.0
